Changed the labels for the Maps table for the FVS_Ststs_Species table which was crashing during rendering due to the recent adding of 3 species to each table.
When hovering over a stand polygon in Maps, the pop up was only displaying the HTML code and not the table. This was due to the list "labs" being populated with single brackets and not double. 

The Delete Project functionality was referencing the name of a recently-replaced variable, hence no projects were ever getting deleted.