FVSOnlineDev: Fixed an issue in Maps where it was not using the output database for the list of runs that could be displayed...also checked on query errors.
Modified "create project" making it clear that it does not copy the runs or
input or output data from the source project and it does not start the new run.
(this change consolidates the project switching code). This change includes
making sure the copy is just the relevent files.

Modified the "onlocal" but not Windows project switching code. Now it starts the
new project in a new browser window. The windows code is unchanged, it closes the 
window when the new window starts.